+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Renamed preset to Windblade.
 - Renamed all size units.
 - Size units are now sorted between `units` and `misc` in theme.